+++ Java/tempctrl/src/main/java/com/easleydp/tempctrl/domain/TemperatureProfileDto.java
@@ -59,7 +59,7 @@ public class TemperatureProfileDto
 
         /** Convenience */
         @JsonIgnore
-        public int getMillisSinceStart()
+        public long getMillisSinceStart()
         {
             return hoursSinceStart * 1000 * 60 * 60;
         }
